Добавили  log4j2

See merge request kata-project1/market-product!8

+++ src/main/java/com/example/demo/services/implementation/ProductServiceImplementation.java
@@ -1,5 +1,6 @@
 package com.example.demo.services.implementation;
 
+import com.example.demo.exception_handling.NoSuchProductException;
 import com.example.demo.models.Product;
 import com.example.demo.repositories.ProductRepository;
 import com.example.demo.services.ProductService;
@@ -17,31 +18,47 @@ public class ProductServiceImplementation implements ProductService {
 
     @Override
     @Transactional
-    public void addProduct(Product product) {
-        productRepository.save(product);
+    public Product addProduct(Product product) {
+        return productRepository.save(product);
     }
 
     @Override
     @Transactional(readOnly = true)
     public List<Product> getAllProducts() {
-        return productRepository.findAll();
+
+        List<Product> products = productRepository.findAll();
+        if ((products == null) || (products.isEmpty())) {
+            throw new NoSuchProductException("No product found in Database");
+        }
+        return products;
     }
 
     @Override
     @Transactional(readOnly = true)
     public Product findProductById(long id) {
-        return productRepository.findProductById(id);
+        Product product = productRepository.findProductById(id);
+        if (product == null) {
+            throw new NoSuchProductException("There is no product with ID = " + id + " in Database");
+        }
+        return product;
     }
 
     @Override
     @Transactional
-    public void updateProduct(Product product) {
-        productRepository.save(product);
+    public Product updateProduct(Product product) {
+        Product oldProduct = productRepository.findProductById(product.getId());
+        if (oldProduct == null) {
+            throw new NoSuchProductException("There is no product with ID = " + product.getId() + " in Database");
+        }
+        return productRepository.save(product);
     }
 
     @Override
     @Transactional
     public void deleteProductById(long id) {
+        if (productRepository.findProductById(id) == null) {
+            throw new NoSuchProductException("There is no product with ID = " + id + " in Database");
+        }
         productRepository.deleteById(id);
     }
 }
